solved inssues in save_m_obb api

+++ app.py
@@ -118,23 +118,31 @@ async def categorize(
     return response
 
 @app.get("/save_m_obb")
-def save_m_obb(response:str = Form(...)):
-    response = json.loads(response)
+def save_m_obb(
+    file_name: str = Form(...),
+    pg_no: int = Form(...),
+    category: str = Form(...),
+):
     save_file = "obb-traindata.json"
+    new_entry = {
+        "file_name": file_name,
+        "pg_no": pg_no,
+        "category": category
+    }
+    
     if os.path.exists(save_file):
         with open(save_file, "r") as json_file:
             existing_data = json.load(json_file)
     else:
         existing_data = []
-    existing_data.append({
-        "file_name": response["file_name"],
-        "pg_no": response["pg_no"],
-        "category": response["category"],
-    })
-    with open(save_file, "w") as json_file:
-        json.dump(existing_data, json_file)
 
-    return {"message": "Data saved successfully"}
+    if new_entry not in existing_data:
+        existing_data.append(new_entry)
+        with open(save_file, "w") as json_file:
+            json.dump(existing_data, json_file)
+        return {"message": "Data saved successfully"}
+    else:
+        return {"message": "Data already exists"}
 
 @app.post("/set_dpi")
 async def set_dpi(
